backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.1.0
celery[redis]>=5.4.0
redis>=5.0.0
Pillow>=10.0.0
qrcode[pil]>=7.4.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
django-storages[boto3]>=1.14.0
django-ratelimit>=4.1.0
dj-database-url>=2.2.0
pytest>=8.2.0
pytest-django>=4.9.0
fedapay>=0.3.0
requests>=2.32.0
sentry-sdk>=2.0.0
websockets>=12.0
